refactor: rename cell boundary to daemon host
