+++ apps/agor-daemon/src/host/availability.ts
