+++ apps/agor-daemon/src/register-services.ts
@@ -48,8 +48,6 @@ import type {
 import { ROLES, TaskStatus } from '@agor/core/types';
 import type { UnixUserMode } from '@agor/core/unix';
 import type express from 'express';
-import { shouldRegisterLocalHostOperations } from './cell/availability.js';
-import { createLocalCellHostOperations } from './cell/local/local-cell-host-operations.js';
 import type {
   BoardsServiceImpl,
   MessagesServiceImpl,
@@ -61,6 +59,8 @@ import {
   markExecutorProcessExited,
   trackExecutorProcess,
 } from './executor-tracking.js';
+import { shouldRegisterLocalHostOperations } from './host/availability.js';
+import { createLocalDaemonHostOperations } from './host/local/local-daemon-host-operations.js';
 import { runInOAuthTenantScope } from './oauth-auth-helpers.js';
 import {
   cacheOAuth21Token,
@@ -554,11 +554,11 @@ export async function registerServices(ctx: RegisterServicesContext): Promise<Re
 
   const configService = createConfigService(db);
   configService.app = app;
-  // Host ACL/user/group operations exist only on a self-hosted Cell. Hosted
+  // Host ACL/user/group operations exist only on a self-hosted daemon host. Hosted
   // registration is intentionally absent rather than forwarding privileged
   // work through an impersonated executor.
   if (shouldRegisterLocalHostOperations(config)) {
-    app.use('/admin/local-actions', createLocalActionsService(createLocalCellHostOperations()));
+    app.use('/admin/local-actions', createLocalActionsService(createLocalDaemonHostOperations()));
   }
 
   app.use('/agentic-tool-settings', createTenantAgenticToolSettingsService(db));
