+++ apps/agor-daemon/src/services/local-actions.ts
@@ -2,7 +2,7 @@ import { BadRequest } from '@agor/core/feathers';
 import type { LocalActionResult } from '@agor/core/local-actions';
 import type { Params } from '@agor/core/types';
 import { AGOR_HOME_BASE } from '@agor/core/unix';
-import type { CellHostOperations } from '../cell/host-operations.js';
+import type { DaemonHostOperations } from '../host/operations.js';
 
 export type LocalActionName =
   | 'unix.group.createBranch'
@@ -33,7 +33,7 @@ function homeBase(params: Record<string, unknown>): string | undefined {
     throw new BadRequest(`homeBase must be the managed Agor home base: ${AGOR_HOME_BASE}`);
   return AGOR_HOME_BASE;
 }
-export function createLocalActionsService(host: CellHostOperations) {
+export function createLocalActionsService(host: DaemonHostOperations) {
   return {
     async create(data: LocalActionRequest, _params?: Params): Promise<LocalActionResult> {
       const p = data.params ?? {};
